+++ curricula/models/common.py
@@ -1,0 +1,36 @@
+from django.db import models
+from django.contrib.contenttypes.models import ContentType
+from django.core import urlresolvers
+
+
+def get_earliest_gap(seq):
+    """
+    Find the earliest gap in `seq` which should be a list of
+    sequential numbers.
+    """
+    for i in range(len(seq) + 1):
+        if i not in seq:
+            return i
+
+
+class BaseModel(models.Model):
+
+    class Meta:
+        abstract = True
+
+    created_on = models.DateTimeField(auto_now_add=True)
+    updated_on = models.DateTimeField(auto_now=True)
+
+    def instance_from_db(self):
+        return self.__class__.objects.get(pk=self.pk)
+
+    def get_admin_url(self):
+        content_type = ContentType.objects.get_for_model(self.__class__)
+        return urlresolvers.reverse(
+            'admin:{}_{}_change'.format(
+                content_type.app_label,
+                content_type.model
+            ),
+            args=[self.id]
+        )
+
